Module indexing::algorithms

Respository of some indexing-implemented algorithms so we can dissect them and their codegen.

Traits

Data
Convenience trait -- debugging

Functions

binary_search

heapify

insertion_sort_indexes

insertion_sort_pointerindex

insertion_sort_ranges

insertion_sort_rust
Copied from rust / libcollections/slice.rs, using raw pointers

merge_internal_indices
Merge internal: Merge inside data while using buffer as a swap space

merge_internal_ranges
Merge internal: Merge inside data while using buffer as a swap space

quicksort
Simple quicksort implemented using indexing,

quicksort_bounds
quicksort implemented using regular bounds checked indexing