Module indexing::container_traits

Traits

Contiguous

The container has a contiguous addressable range.

ContiguousMut

FixedLength

The container does not change is length while we are trusting it

GetUnchecked

GetUncheckedMut

IndexRange

A range being .., a.., ..b, or a..b.

OnePointRange

A range with at most one point, being a.., .., or ..b.

Pushable

Trustworthy

The most basic container trait: it can have indices and ranges that are trusted to be in bounds.