Struct indexing::pointer::PRange

pub struct PRange<'id, T, Proof = Unknown> { /* private fields */ }

PRange is a pointer-based valid range with start and end pointer representation.

Implementations

impl<'id, T, P> PRange<'id, T, P>

pub fn len(self) -> usize

pub fn is_empty(self) -> bool

pub fn nonempty(&self) -> Result<PRange<'id, T, NonEmpty>, IndexingError>

Check if the range is empty. NonEmpty ranges have extra methods.

pub fn split_in_half(self) -> (PRange<'id, T>, PRange<'id, T, P>)

Split the range in half, with the upper middle index landing in the latter half. Proof of nonemptiness P transfers to the latter half.

impl<'id, T, P> PRange<'id, T, P>

pub fn first(self) -> PIndex<'id, T, P>

pub fn upper_middle(self) -> PIndex<'id, T, P>

Return the middle index, rounding up on even

pub fn past_the_end(self) -> PIndex<'id, T, Unknown>

impl<'id, T> PRange<'id, T, NonEmpty>

pub fn last(self) -> PIndex<'id, T>

pub fn tail(self) -> PRange<'id, T>

pub fn init(self) -> PRange<'id, T>

pub fn advance(&mut self) -> bool

Increase the range’s start, if the result is still a non-empty range.

Return true if stepped successfully, false if the range would be empty.

pub fn advance_back(&mut self) -> bool

Decrease the range’s end, if the result is still a non-empty range.

Return true if stepped successfully, false if the range would be empty.