Struct indexing::Index

pub struct Index<'id, Proof = NonEmpty> { /* private fields */ }

A branded index.

Index<'id> only indexes the container instantiated with the exact same particular lifetime for the parameter 'id at its inception from the scope() function.

The type parameter Proof determines if the index is dereferenceable.

A NonEmpty index points to a valid element. An Unknown index is unknown, or it points to an edge index (just past the end).

Implementations

impl<'id, P> Index<'id, P>

pub fn integer(&self) -> usize

Return the index as an integer offset from the start of the container

impl<'id> Index<'id, NonEmpty>

pub fn after(self) -> Index<'id, Unknown>

Return the index directly after.